+++ src/mailmanclient/docs/NEWS.rst
@@ -7,7 +7,9 @@ NEWS for mailmanclient
 
 3.3.3 (2021-XX-XX)
 ==================
-- Add ``pre_confirmed`` and ``pre_approved`` parameters to ``MailingList.unsubscribe`` (Fixes #62)
+- Add ``pre_confirmed`` and ``pre_approved`` parameters to
+  ``MailingList.unsubscribe``. (Fixes #62)
+- Add support to fetch pending unsubscription requests. (Cloes #63)
 
 .. _news-3.3.2: