+++ src/mailmanclient/docs/NEWS.rst
@@ -9,7 +9,7 @@ NEWS for mailmanclient
 ==================
 - Add ``pre_confirmed`` and ``pre_approved`` parameters to
   ``MailingList.unsubscribe``. (Fixes #62)
-- Add support to fetch pending unsubscription requests. (Cloes #63)
+- Add support to fetch pending unsubscription requests. (Closes #63)
 
 .. _news-3.3.2: