+++ src/mailmanclient/docs/NEWS.rst
@@ -10,6 +10,7 @@ NEWS for mailmanclient
 - Add ``pre_confirmed`` and ``pre_approved`` parameters to
   ``MailingList.unsubscribe``. (Fixes #62)
 - Add support to fetch pending unsubscription requests. (Closes #63)
+- Add ``member_id`` as a property of ``Member`` object. (Closes #64)
 
 .. _news-3.3.2: